modification of cli desc

+++ x/token/client/cli/query.go
@@ -40,7 +40,7 @@ func GetQueryCmd(queryRoute string) *cobra.Command {
 func GetCmdQueryToken() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "info [denom]",
-		Long:    "Query a token detail by symbol or min unit.",
+		Short:   "Query a token detail by symbol or min unit.",
 		Example: fmt.Sprintf("$ %s query token info <denom>", version.AppName),
 		Args:    cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
@@ -70,7 +70,7 @@ func GetCmdQueryToken() *cobra.Command {
 func GetCmdQueryParams() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "params",
-		Long:    "Query values set as token parameters",
+		Short:   "Query values set as token parameters",
 		Example: fmt.Sprintf("$ %s q token params", version.AppName),
 		Args:    cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, args []string) error {
@@ -94,7 +94,7 @@ func GetCmdQueryParams() *cobra.Command {
 func GetCmdQueryTokens() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "tokens [owner]",
-		Long:    "Query tokens by the owner",
+		Short:   "Query tokens by the owner",
 		Example: fmt.Sprintf("$ %s q token tokens <owner>", version.AppName),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx, err := client.GetClientQueryContext(cmd)
@@ -143,7 +143,7 @@ func GetCmdQueryTokens() *cobra.Command {
 func GetCmdQueryTokenFee() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "fee [symbol]",
-		Long:    "Query the token related fees",
+		Short:   "Query the token related fees",
 		Args:    cobra.ExactArgs(1),
 		Example: fmt.Sprintf("$ %s q token fee <symbol>", version.AppName),
 		RunE: func(cmd *cobra.Command, args []string) error {
